refactor: move files to they own folders
