+++ service/flooding/FloodingAlgorithm.java
@@ -1,8 +1,8 @@
-package service;
+package service.flooding;
 
 import model.Packet;
 import model.Router;
 
-public interface IController {
+public interface FloodingAlgorithm {
   public void send(Router from, Router router, Packet packet);
 }
